test(wadjet): fix stale assertions after #530 int4 and #536 int-literal refusal
The ./wadjet/ package was not in #530/#536/#579's gate battery nor in the
pre-push hook (which runs ./internal/... only), so three landed correctness
fixes left it red:

- #530 made LENGTH/POSITION/octet_length/etc. return int4 (was float8);
  pg_client_shapes_test.go, postgres_gaps_test.go, and
  generic_string_func_network_args_test.go still asserted float64 (values
  already correct). Updated to int32.
- #536 correctly refuses a non-numeric literal against an integer column
  (22P02); decimal_literal_test.go's 'still answers' set listed
  'WHERE k = 'abc'' (k bigint) as a should-answer case — now correctly
  refused. Dropped it (covered by filter_int_literal_guard_test.go).

Values were verified correct; only the type/refusal expectations were stale.

+++ wadjet/generic_string_func_network_args_test.go
@@ -28,7 +28,7 @@ func TestGenericStringFuncsRenderNetworkColumnArgumentsAsText(t *testing.T) {
 		want any
 	}{
 		// --- the two types that box as a raw int64 (the reported bug) ---
-		{"length(ipv4)", `SELECT length(c_ipv4) AS v FROM net_matrix WHERE id = 1`, float64(len("10.1.2.3"))},
+		{"length(ipv4)", `SELECT length(c_ipv4) AS v FROM net_matrix WHERE id = 1`, int32(len("10.1.2.3"))},
 		{"concat(ipv4, '!')", `SELECT concat(c_ipv4, '!') AS v FROM net_matrix WHERE id = 1`, "10.1.2.3!"},
 		{"upper(mac)", `SELECT upper(c_mac) AS v FROM net_matrix WHERE id = 1`, "AA:BB:CC:DD:EE:FF"},
 		{"lower(mac)", `SELECT lower(c_mac) AS v FROM net_matrix WHERE id = 1`, "aa:bb:cc:dd:ee:ff"},
@@ -42,10 +42,10 @@ func TestGenericStringFuncsRenderNetworkColumnArgumentsAsText(t *testing.T) {
 		// --- already-correct types, pinned so a future change can't break
 		// them while "fixing" IPv4/MAC (same regression-guard shape as
 		// network_typed_column_args_test.go) ---
-		{"length(ipv6) (guard)", `SELECT length(c_ipv6) AS v FROM net_matrix WHERE id = 1`, float64(len("2001:db8::1"))},
-		{"length(cidr) (guard)", `SELECT length(c_cidr) AS v FROM net_matrix WHERE id = 1`, float64(len("192.168.1.0/24"))},
-		{"length(port) (guard)", `SELECT length(c_port) AS v FROM net_matrix WHERE id = 1`, float64(len("443"))},
-		{"length(proto) (guard)", `SELECT length(c_proto) AS v FROM net_matrix WHERE id = 1`, float64(len("6"))},
+		{"length(ipv6) (guard)", `SELECT length(c_ipv6) AS v FROM net_matrix WHERE id = 1`, int32(len("2001:db8::1"))},
+		{"length(cidr) (guard)", `SELECT length(c_cidr) AS v FROM net_matrix WHERE id = 1`, int32(len("192.168.1.0/24"))},
+		{"length(port) (guard)", `SELECT length(c_port) AS v FROM net_matrix WHERE id = 1`, int32(len("443"))},
+		{"length(proto) (guard)", `SELECT length(c_proto) AS v FROM net_matrix WHERE id = 1`, int32(len("6"))},
 	}
 
 	for _, tc := range cases {
@@ -80,11 +80,11 @@ func TestGenericStringFuncsRenderNetworkColumnArgumentsAsTextVectorized(t *testi
 		t.Fatalf("query failed: %v", err)
 	}
 	want := []struct {
-		l float64
+		l int32
 		u string
 	}{
-		{float64(len("10.1.2.3")), "AA:BB:CC:DD:EE:FF"},
-		{float64(len("192.168.1.1")), "11:22:33:44:55:66"},
+		{int32(len("10.1.2.3")), "AA:BB:CC:DD:EE:FF"},
+		{int32(len("192.168.1.1")), "11:22:33:44:55:66"},
 	}
 	if len(res.Rows) != len(want) {
 		t.Fatalf("got %d rows, want %d", len(res.Rows), len(want))
